Chore(verify) Move verification code to email controller
Creating a separate file in the root folder of the application to verify
a user's email (when they click the link) is unnecessary and could all
be done in the email controller which exists for the purpose. Therefore
move verify_email.php contents to a function in email controller and
call that function from index.php. verify_email.php can now be safely
deleted.